requirements.txt
streamlit
Pillow
boto3
python-dotenv
openai
faiss-cpu
numpy 